Support expires_in as a number or string
Attempting to get a refresh_token from login.microsoft.com would result
in the following error message.

    failed to parse exchange response: json: cannot unmarshal string into Go struct field TokenResponse.expires_in of type int64

This happens because their API is returning `expires_in` as a JSON
string instead of as a JSON number as expected. It seems this is in
violation of the OAuth2 spec, but it's happening nonetheless.

This change allows `expires_in` to be accepted as an int64 whether it's
encoded as a JSON number or string.